requirements.txt
requests
beautifulsoup4
pandas
python-binance
supabase
python-dotenv